+++ src/main/java/lotto/common/LottoPriceInfo.java
@@ -1,4 +1,4 @@
-package lotto;
+package lotto.common;
 
 import java.util.Arrays;
 import java.util.Map;
