+++ src/main/java/lotto/common/LottoShuffleMachine.java
@@ -1,4 +1,4 @@
-package lotto;
+package lotto.common;
 
 import java.util.Collections;
 import java.util.HashSet;
@@ -6,6 +6,7 @@ import java.util.List;
 import java.util.Set;
 import java.util.stream.Collectors;
 import java.util.stream.IntStream;
+import lotto.domain.LottoNumber;
 
 public class LottoShuffleMachine {
