+++ src/main/java/lotto/common/LottoTicketMachine.java
@@ -1,10 +1,12 @@
-package lotto;
+package lotto.common;
 
 import java.util.Arrays;
 import java.util.List;
 import java.util.Set;
 import java.util.stream.Collectors;
 import java.util.stream.Stream;
+import lotto.domain.LottoNumber;
+import lotto.domain.LottoTicket;
 
 public class LottoTicketMachine {
