+++ src/test/java/lotto/common/LottoShuffleMachineTest.java
@@ -1,8 +1,9 @@
-package lotto;
+package lotto.common;
 
 import static org.assertj.core.api.Assertions.assertThat;
 
 import java.util.Set;
+import lotto.domain.LottoNumber;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
